pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-rag-diploma"
version = "0.1.0"
description = "Model-agnostic enterprise cloud RAG onboarding chatbot proof of concept"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12,<5.0",
    "fastapi>=0.115,<1.0",
    "httpx>=0.27,<1.0",
    "msal>=1.31,<2.0",
    "opentelemetry-api>=1.27,<2.0",
    "opentelemetry-exporter-otlp>=1.27,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0,<1.0",
    "opentelemetry-instrumentation-httpx>=0.48b0,<1.0",
    "opentelemetry-sdk>=1.27,<2.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "PyJWT[crypto]>=2.9,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "pypdf>=5.4,<6.0",
    "qdrant-client>=1.10,<2.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]
perf = [
    "locust>=2.32,<3.0",
]

[project.scripts]
rag-api = "rag_api.main:run"
sync-worker = "sync_worker.main:run"
sharepoint_bootstrap = "sync_worker.jobs.sharepoint_bootstrap:run"
sharepoint_incremental = "sync_worker.jobs.sharepoint_incremental:run"
onenote_bootstrap = "sync_worker.jobs.onenote_bootstrap:run"
onenote_incremental = "sync_worker.jobs.onenote_incremental:run"
ops_worker = "sync_worker.jobs.ops_worker:run"
subscription_renewal = "sync_worker.jobs.subscription_renewal:run"
sharepoint_delta_catchup = "sync_worker.jobs.sharepoint_delta_catchup:run"
sharepoint_reconciliation = "sync_worker.jobs.sharepoint_reconciliation:run"
onenote_reconciliation = "sync_worker.jobs.onenote_reconciliation:run"
rag_evaluate = "rag_api.evaluation.cli:run"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = [
    "services/rag-api/app",
    "services/sync-worker/app",
    "libs/shared-schemas/python",
    "services/graph-connectors",
]
include = [
    "graph_connectors*",
    "rag_api*",
    "shared_schemas*",
    "sync_worker*",
]
